Added the removemanga command and mangalist
Now you can remove manga from the alert list using the bot itself!
You have to insert the exact match that is in the alert list, but don't worry, if you get it wrong the bot will tell you the closest matches. 
You can also ask for the complete manga list, so you can see which manga you are following!